# perf(tm2/pkg/amino): reduce RAM heavy-handedness by *bytes.Buffer pooled reuse