--- a/tm2/pkg/amino/codec.go
+++ b/tm2/pkg/amino/codec.go
@@ -1,7 +1,6 @@
 package amino
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 	"reflect"
@@ -113,7 +112,12 @@ func (info *TypeInfo) String() string {
 		// before it's fully populated.
 		return "<new TypeInfo>"
 	}
-	buf := new(bytes.Buffer)
+	buf := poolBytesBuffer.Get()
+	defer func() {
+		buf.Reset()
+		poolBytesBuffer.Put(buf)
+	}()
+
 	buf.Write([]byte("TypeInfo{"))
 	buf.Write([]byte(fmt.Sprintf("Type:%v,", info.Type)))
 	if info.ConcreteInfo.Registered {