--- a/tm2/pkg/amino/json_encode.go
+++ b/tm2/pkg/amino/json_encode.go
@@ -1,7 +1,6 @@
 package amino
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -156,7 +155,12 @@ func (cdc *Codec) encodeReflectJSONInterface(w io.Writer, iinfo *TypeInfo, rv re
 	}
 
 	// Write Value to buffer
-	buf := new(bytes.Buffer)
+	buf := poolBytesBuffer.Get()
+	defer func() {
+		buf.Reset()
+		poolBytesBuffer.Put(buf)
+	}()
+
 	cdc.encodeReflectJSON(buf, cinfo, crv, fopts)
 	value := buf.Bytes()
 	if len(value) == 0 {