@odeke-em 2025-01-13
Hey @thehowl I am would not advise using libp2p-go-pool's Buffer and here are some benchmark results from their own benchmarks that I added the Go standard library bytes.Buffer benchmarks https://github.com/orijtech/go-buffer-pool/commit/8ea7d61d18c86b7eab3ab111e180db92ff1845ab 

```shell
$ benchstat libp2p.txt stdlib.txt
name                       old time/op    new time/op    delta
WriteByte-8                  16.8µs ± 4%    16.7µs ± 9%      ~     (p=0.400 n=9+10)
BufferFullSmallReads-8       61.1µs ± 6%    46.4µs ± 3%   -24.06%  (p=0.000 n=10+10)
BufferNotEmptyWriteRead-8     417µs ± 3%     182µs ± 5%   -56.28%  (p=0.000 n=10+10)

name                       old speed      new speed      delta
WriteByte-8                 244MB/s ± 3%   245MB/s ± 8%      ~     (p=0.400 n=9+10)

name                       old alloc/op   new alloc/op   delta
WriteByte-8                   0.00B          0.00B           ~     (all equal)
BufferFullSmallReads-8        0.00B          0.00B           ~     (all equal)
BufferNotEmptyWriteRead-8    0.90B ±122%     0.00B       -100.00%  (p=0.011 n=10+10)

name                       old allocs/op  new allocs/op  delta
WriteByte-8                    0.00           0.00           ~     (all equal)
BufferFullSmallReads-8         0.00           0.00           ~     (all equal)
BufferNotEmptyWriteRead-8      0.00           0.00           ~     (all equal)
```
and that's just for the Buffer; the standard library's sync.Pool is battle tested across millions of repositories and is continually being improved while libp2p doesn't get much attention