@odeke-em 2025-01-14
Thanks for the response @thehowl!
 
>My issue is as follows: using a byte buffer with a sync.Pool directly is not useful, as the buffer you get out may have any size of the underlying buffer, which is not guaranteed to be reused. Some buffers may become very very large and end up wasting memory.

Before being put back into the pool, we invoke buf.Reset() which then internally moves the underlying byteslice for reuse with b.buf[:0] by making it empty, but its capacity won't have to be re-created, which is where the savings come from. I believe that alleviates your concern. It is directly useful as the incumbent pool used in many high performance projects too.

>Here's another suggestion: https://github.com/valyala/bytebufferpool - it seems to implement the kind of protection I was worried about :)

I have more faith in valyala's implementation, thank you! Sadly though, it doesn't fully implement *bytes.Buffer's methods and in some instances like [json.Indent](https://pkg.go.dev/encoding/json#Indent) which is used inside amino in more than 3 places  which takes a *bytes.Buffer directly.