@thehowl 2025-01-14
> Before being put back into the pool, we invoke buf.Reset() which then internally moves the underlying byteslice for reuse with b.buf[:0] by making it empty, but its capacity won't have to be re-created, which is where the savings come from. I believe that alleviates your concern. It is directly useful as the incumbent pool used in many high performance projects too.

yes, but if I'm encoding a value which takes 16MB to store (which would happen quite exceptionally, anyway) I want the allocated buffer to be freed rather than being put into the buffer. So, either we hard-code a limit beyond which we _don't_ put things back into the buffer, or we use something like bytebufferpool (with the additional methods) which automatically calibrates to an ideal amount of buffer sizes